requirements.txt
# ~/WikiGraph/requirements.txt

# Core parsing
mwxml>=0.3.0
mwparserfromhell>=0.6.4

# Data processing
pandas>=2.0.0
tqdm>=4.65.0
ujson>=5.8.0
pyyaml>=6.0.0

# Database
neo4j>=5.14.0
networkx>=3.0
sqlalchemy>=2.0.0

# Web framework
flask>=3.0.0
flask-cors>=4.0.0

# Development tools
memory-profiler>=0.61.0
psutil>=5.9.0
